Fixed RxJS uses

+++ frontend/src/app/profile/services/profile-delete.service.ts
@@ -3,15 +3,17 @@ import {NotificationService} from '../../core/services/notification.service';
 import {NavigatorService} from '../../core/services/navigator.service';
 import {PlayerApiService} from '../../core/api/services/player-api.service';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable()
 export class ProfileDeleteService {
   private readonly playerApiService = inject(PlayerApiService);
   private readonly notificationService = inject(NotificationService);
   private readonly navigatorService = inject(NavigatorService);
 
-  deleteProfile(playerId: string): void {
+  deleteProfile(playerId: string | null): void {
+    if (playerId === null) {
+      this.notificationService.error('Unable to delete profile');
+      return;
+    }
     this.playerApiService.deletePlayer(playerId)
       .subscribe({
         next: () => {
